refactor(jina): 모듈명 변경 #52
