+++ src/features/musicPreference/api/useGetMood.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 
 import URL from 'shared/config/urls';
-import axiosInstance from 'shared/hooks/useAxiosInstance';
+import axiosInstance from 'shared/lib/axiosInstance';
 
 // 메인데이터 호출
 const GetMood = async () => {
